.bashrc: Add additional checks for the `.bashrc_scripts`
Previously, the `.bashrc` would fail, if the `.bashrc_scripts`
directory was missing. We made sure it existed, but now it would
fail, if there was no `.sh`-script inside the directory.

This commit fixes that and furthermore adds a check for scripts
that do not have the correct permissions, in order to make sure
the `.bashrc` is failsave on other machines.