Add infrastructure for rust scripts
This commit adds infrastructure to include rust executables in my
dotfiles. The reason for this is that I want to learn to code in
rust, but need some motivation to do it. Automating my daily tasks
seems like the perfect opportunity for this.

The commit does the following:
- Add a `rust_scripts` directory to store my rust executables
- Adapt the `.bashrc` to include a new `~/.rust_scripts` directory to
  the `$PATH`
- Adapt the `deploy_scripts/deploy_debian_headless.sh` to symlink
  the `rust_scripts` directory to `~/.rust_scripts`.

Furthermore, I replaced the shell script
`.bashrc_scripts/url_to_markdown.sh` with a rust script that does the
same (but better, with a few more options). For that, we also adapt
the alias definitions in the `.bashrc_scripts/alias.sh`.