.bashrc_scripts/date_to_name.sh: Fix datetime_to_filename method
Previously, the `datetime_to_filename` method missed to add
seconds to the filename, which could lead to duplicate times, for
example when having taken several pictures in one minute.

This adds the seconds to the filename and furthemore removes the
colons between the hour, minute and seconds of the ISO-8601 format.
While this is not exactly standard-conform, it is the only way the
filenames are allowed on Windows machines, which I sadly cannot
get around fully at this moment.

Fixes #13